Got the track-following VELO Pixel tracking algorithm to work within the GPU Manager.

+++ Brunel_v47r2p1/Pr/PrPixel/src/PrPixelTypes.h
@@ -23,7 +23,7 @@ struct PixelEvent {
   std::vector<int32_t> hitIDs;
   std::vector<float>   hitXs;
   std::vector<float>   hitYs;
-  std::vector<int32_t> hitZs;
+  std::vector<float>   hitZs;
 };
 
 #endif
